+++ common.tf
@@ -154,9 +154,23 @@ data "aws_iam_policy_document" "clumio_drift_detect_policy_document" {
       "events:DescribeRule",
       "events:ListTargetsByRule"
     ]
-    effect    = "Allow"
-    resources = compact([aws_iam_role.clumio_iam_role.arn, aws_sns_topic.clumio_event_pub.arn, local.should_create_tag_event_rule ? aws_cloudwatch_event_rule.clumio_tag_event_rule[0].arn : "", aws_cloudwatch_event_rule.clumio_aws_backup_event_rule[0].arn, var.is_rds_enabled ? aws_cloudwatch_event_rule.clumio_rds_cloudwatch_event_rule[0].arn : "", var.is_rds_enabled ? aws_cloudwatch_event_rule.clumio_rds_cloudtrail_event_rule[0].arn : "", var.is_dynamodb_enabled ? aws_cloudwatch_event_rule.clumio_dynamo_cloudtrail_event_rule[0].arn : "", var.is_ebs_enabled ? aws_cloudwatch_event_rule.clumio_ebs_cloudwatch_event_rule[0].arn : "", var.is_ebs_enabled ? aws_cloudwatch_event_rule.clumio_ec2_cloudwatch_event_rule[0].arn : "", var.is_ebs_enabled ? aws_cloudwatch_event_rule.clumio_ebs_cloudtrail_event_rule[0].arn : "", var.is_ebs_enabled ? aws_cloudwatch_event_rule.clumio_ec2_cloudtrail_event_rule[0].arn : ""])
-    sid       = "ReflectOnClumioCfnStack"
+    effect = "Allow"
+    resources = compact([aws_iam_role.clumio_iam_role.arn, aws_sns_topic.clumio_event_pub.arn,
+      local.should_create_tag_event_rule ? aws_cloudwatch_event_rule.clumio_tag_event_rule[0].arn : "",
+      var.is_rds_enabled ? aws_cloudwatch_event_rule.clumio_rds_cloudwatch_event_rule[0].arn : "",
+      var.is_rds_enabled ? aws_cloudwatch_event_rule.clumio_rds_aws_backup_cloudwatch_event_rule[0].arn : "",
+      var.is_rds_enabled ? aws_cloudwatch_event_rule.clumio_rds_cloudtrail_event_rule[0].arn : "",
+      var.is_dynamodb_enabled ? aws_cloudwatch_event_rule.clumio_dynamo_cloudtrail_event_rule[0].arn : "",
+      var.is_dynamodb_enabled ? aws_cloudwatch_event_rule.clumio_dynamo_aws_backup_cloudwatch_event_rule[0].arn : "",
+      var.is_ebs_enabled ? aws_cloudwatch_event_rule.clumio_ebs_cloudwatch_event_rule[0].arn : "",
+      var.is_ebs_enabled ? aws_cloudwatch_event_rule.clumio_ec2_cloudwatch_event_rule[0].arn : "",
+      var.is_ebs_enabled ? aws_cloudwatch_event_rule.clumio_ec2_aws_backup_cloudwatch_event_rule[0].arn : "",
+      var.is_ebs_enabled ? aws_cloudwatch_event_rule.clumio_ebs_aws_backup_cloudwatch_event_rule[0].arn : "",
+      var.is_ebs_enabled ? aws_cloudwatch_event_rule.clumio_ebs_cloudtrail_event_rule[0].arn : "",
+      var.is_ebs_enabled ? aws_cloudwatch_event_rule.clumio_ec2_cloudtrail_event_rule[0].arn : "",
+      var.is_s3_enabled ? aws_cloudwatch_event_rule.clumio_s3_cloudtrail_event_rule[0].arn : "",
+    var.is_s3_enabled ? aws_cloudwatch_event_rule.clumio_s3_aws_backup_cloudwatch_event_rule[0].arn : "", ])
+    sid = "ReflectOnClumioCfnStack"
   }
 }
 
@@ -294,7 +308,7 @@ data "aws_iam_policy_document" "clumio_inventory_policy_document" {
     }
   }
 
-  # Allow Clumio to get Recovery Point info
+  # Allow Clumio to get AWS Recovery Point info
   dynamic "statement" {
     for_each = [1]
     content {
@@ -303,12 +317,7 @@ data "aws_iam_policy_document" "clumio_inventory_policy_document" {
       ]
       effect = "Allow"
       resources = [
-        "arn:${data.aws_partition.current.partition}:ec2:${var.aws_region}:*:snapshot/*",
-        "arn:${data.aws_partition.current.partition}:ec2:${var.aws_region}:*:image/*",
-        "arn:${data.aws_partition.current.partition}:dynamodb:${var.aws_region}:${var.aws_account_id}:table/*",
-        "arn:${data.aws_partition.current.partition}:rds:${var.aws_region}:${var.aws_account_id}:snapshot:*",
-        "arn:${data.aws_partition.current.partition}:rds:${var.aws_region}:${var.aws_account_id}:cluster-snapshot:*",
-        "arn:${data.aws_partition.current.partition}:backup:${var.aws_region}:${var.aws_account_id}:recovery-point:*"
+        "*"
       ]
       sid = "GetAWSRecoveryPointInfo"
     }
@@ -822,23 +831,6 @@ resource "aws_cloudwatch_event_target" "clumio_tag_event_rule_target" {
   target_id = "clumio-publish"
 }
 
-resource "aws_cloudwatch_event_rule" "clumio_aws_backup_event_rule" {
-  count = 1
-  depends_on = [
-    time_sleep.wait_before_create
-  ]
-  description   = "Watches for AWS backup changes"
-  event_pattern = "{\"source\": [\"aws.backup\"], \"detail-type\": [\"Recovery Point State Change\"], \"detail\": {\"resourceType\": [\"Aurora\", \"DynamoDB\", \"DynamoDB.FullyManaged\", \"EBS\", \"EC2\", \"RDS\", \"RDS.Cluster\", \"S3\"], \"status\": [\"COMPLETED\", \"PARTIAL\", \"EXPIRED\", \"FAILED\", \"DELETED\"]}}"
-  name          = "ClumioBackupCloudwatchRule_${var.clumio_token}"
-}
-
-resource "aws_cloudwatch_event_target" "clumio_aws_backup_event_rule_target" {
-  count     = 1
-  arn       = aws_sns_topic.clumio_event_pub.arn
-  rule      = aws_cloudwatch_event_rule.clumio_aws_backup_event_rule[0].name
-  target_id = "clumio-publish"
-}
-
 # The base Clumio policy
 resource "aws_iam_policy" "clumio_base_managed_policy" {
   count  = 1
@@ -941,7 +933,6 @@ resource "clumio_post_process_aws_connection" "clumio_callback" {
     aws_iam_role_policy.clumio_inventory_policy,
     aws_iam_role_policy.clumio_kms_policy,
     aws_cloudwatch_event_target.clumio_tag_event_rule_target,
-    aws_cloudwatch_event_target.clumio_aws_backup_event_rule_target,
     aws_iam_policy.clumio_ec2_mssql_backup_restore_policy,
     aws_iam_instance_profile.clumio_ec2_mssql_ssm_instance_profile,
     aws_iam_role.clumio_ec2_mssql_ssm_instance_role_v2,
@@ -967,17 +958,22 @@ resource "clumio_post_process_aws_connection" "clumio_callback" {
     aws_iam_policy.clumio_s3_backup_policy,
     aws_iam_policy.clumio_s3_restore_policy,
     aws_cloudwatch_event_target.clumio_s3_cloudtrail_event_rule_target,
+    aws_cloudwatch_event_target.clumio_s3_aws_backup_cloudwatch_event_rule_target,
     aws_iam_role.clumio_s3_continuous_backup_event_bridge_role,
     aws_iam_policy.clumio_s3_continuous_backup_event_bridge_policy,
     aws_iam_policy.clumio_dynamodb_backup_policy,
     aws_iam_policy.clumio_dynamodb_restore_policy,
     aws_cloudwatch_event_target.clumio_dynamo_cloudtrail_event_rule_target,
+    aws_cloudwatch_event_target.clumio_dynamo_aws_backup_cloudwatch_event_rule_target,
     aws_cloudwatch_event_target.clumio_ebs_cloudwatch_event_rule_target,
     aws_cloudwatch_event_target.clumio_ebs_cloudtrail_event_rule_target,
+    aws_cloudwatch_event_target.clumio_ebs_aws_backup_cloudwatch_event_rule_target,
     aws_cloudwatch_event_target.clumio_ec2_cloudwatch_event_rule_target,
     aws_cloudwatch_event_target.clumio_ec2_cloudtrail_event_rule_target,
+    aws_cloudwatch_event_target.clumio_ec2_aws_backup_cloudwatch_event_rule_target,
     aws_cloudwatch_event_target.clumio_rds_cloudwatch_event_rule_target,
-    aws_cloudwatch_event_target.clumio_rds_cloudtrail_event_rule_target
+    aws_cloudwatch_event_target.clumio_rds_cloudtrail_event_rule_target,
+    aws_cloudwatch_event_target.clumio_rds_aws_backup_cloudwatch_event_rule_target
   ]
   discover_version      = "4.6"
   intermediate_role_arn = "arn:aws:iam::${var.clumio_aws_account_id}:role/ClumioCustomerProtectRole"
@@ -990,13 +986,13 @@ resource "clumio_post_process_aws_connection" "clumio_callback" {
     "CreateClumioInventoryTopicEncryptionKey" : var.create_clumio_inventory_sns_topic_encryption_key,
     "ClumioInventoryTopicEncryptionKey" : var.clumio_inventory_sns_topic_encryption_key
   }
-  protect_config_version             = "24.1"
+  protect_config_version             = "24.2"
   protect_dynamodb_version           = var.is_dynamodb_enabled ? "7.2" : ""
-  protect_ebs_version                = var.is_ebs_enabled ? "25.0" : ""
+  protect_ebs_version                = var.is_ebs_enabled ? "25.1" : ""
   protect_ec2_mssql_version          = var.is_ec2_mssql_enabled ? "4.4" : ""
-  protect_rds_version                = var.is_rds_enabled ? "20.3" : ""
-  protect_s3_version                 = var.is_s3_enabled ? "7.0" : ""
-  protect_warm_tier_dynamodb_version = var.is_dynamodb_enabled ? "6.0" : ""
+  protect_rds_version                = var.is_rds_enabled ? "21.0" : ""
+  protect_s3_version                 = var.is_s3_enabled ? "7.1" : ""
+  protect_warm_tier_dynamodb_version = var.is_dynamodb_enabled ? "6.1" : ""
   protect_warm_tier_version          = var.is_dynamodb_enabled ? "1.1" : ""
   region                             = var.aws_region
   role_arn                           = aws_iam_role.clumio_iam_role.arn