+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-## 0.27.1
-Changed the Clumio provider version required to >=0.10.0, <0.12.0
+## 0.28.0
+Added data source specific AWS backup cloudwatch event rules. 
+Changed the Clumio provider version required to >=0.10.0, <0.12.0.
+
 
 ## 0.27.0
 New cloudwatch_event_rule added for aws backups.